cuda: reimplement events API support for component
The cuda component this commit replaces was monolitic, with switchs to
select between the CUPTI event/metric API and the profiler API. This
commit introduces a dispatching layer (cuptid) for doing the switching
between the two CUPTI APIs and implements the event/metric support in
the form of a dedicated backend (cuptie) interface. The same is true for
the profiler API, which is refactored in a different commit.