thread: unguard _papi_hwi_my_thread declaration
The _papi_hwi_my_thread variable is defined in threads.c and is made
available to other parts of the code that use it as extern. However, the
extern declaration in both threads.c and threads.h is unnecessarily
guarded. If TLS is not active the variable won't be declared, causing
undefined undeclared errors. Removing the guard from the declaration
should fix the problem.